test(app): colocate port integration
