+++ src/app.rs
@@ -276,153 +276,3 @@ impl App<'_> {
         self.open_status_contact.clone()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app::notifications::notification_is_muted;
-    use crate::app::presence::PresenceMarker;
-    use crate::app::test_support::{TestApp, message};
-    use std::time::Duration;
-
-    #[derive(Debug)]
-    struct FixedClock(Option<i64>);
-
-    impl FixedClock {
-        fn new(value: i64) -> Self {
-            Self(Some(value))
-        }
-    }
-
-    impl Clock for FixedClock {
-        fn unix_seconds(&self) -> Option<i64> {
-            self.0
-        }
-    }
-
-    #[derive(Clone)]
-    struct MutableClock(Arc<Mutex<Option<i64>>>);
-
-    impl MutableClock {
-        fn new(value: Option<i64>) -> Self {
-            Self(Arc::new(Mutex::new(value)))
-        }
-
-        fn set(&self, value: Option<i64>) {
-            *self.0.lock().unwrap() = value;
-        }
-    }
-
-    impl Clock for MutableClock {
-        fn unix_seconds(&self) -> Option<i64> {
-            *self.0.lock().unwrap()
-        }
-    }
-
-    #[derive(Clone, Default)]
-    struct RecordingNotifier {
-        notifications: Arc<Mutex<Vec<(String, String)>>>,
-    }
-
-    impl Notifier for RecordingNotifier {
-        fn show(&self, notification: &NotificationProjection) -> Result<(), String> {
-            self.notifications
-                .lock()
-                .unwrap()
-                .push((notification.summary.to_string(), notification.body.clone()));
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn injected_clock_preserves_mute_boundary_and_presence_timing() {
-        let chat = wr::JID::from("alice@s.whatsapp.net".to_owned());
-        let clock = MutableClock::new(Some(1_000));
-        let mut app = TestApp::with_ports(clock.clone(), RecordingNotifier::default());
-        app.open_chat = Some(chat.clone());
-        let now = app.now();
-        app.selected_presence.select(Some(chat.clone()), now);
-        app.selected_presence.update(&chat, true, 0, now);
-
-        assert!(!notification_is_muted(true, 1_000, app.now()));
-        let now = app.now();
-        assert_eq!(
-            app.selected_presence.marker(Some(&chat), now),
-            Some(PresenceMarker::RecentlyOffline)
-        );
-        assert_eq!(
-            app.selected_presence.redraw_after(app.now()),
-            Some(Duration::from_secs(300))
-        );
-
-        clock.set(Some(1_299));
-        let now = app.now();
-        assert_eq!(
-            app.selected_presence.marker(Some(&chat), now),
-            Some(PresenceMarker::RecentlyOffline)
-        );
-        assert_eq!(
-            app.selected_presence.redraw_after(app.now()),
-            Some(Duration::from_secs(1))
-        );
-        clock.set(Some(1_300));
-        assert!(notification_is_muted(true, 1_301, app.now()));
-        assert!(!notification_is_muted(true, 1_300, app.now()));
-        let now = app.now();
-        assert_eq!(
-            app.selected_presence.marker(Some(&chat), now),
-            Some(PresenceMarker::Offline)
-        );
-        assert_eq!(app.selected_presence.redraw_after(app.now()), None);
-
-        clock.set(None);
-        assert_eq!(app.now(), 0);
-    }
-
-    #[test]
-    fn injected_clock_reaches_presence_and_ui_marker_path() {
-        use ratatui::{Terminal, backend::TestBackend, layout::Rect};
-
-        let chat = wr::JID::from("alice@s.whatsapp.net".to_owned());
-        let mut app =
-            TestApp::with_ports(FixedClock::new(1_700_000_000), RecordingNotifier::default());
-        app.contacts.insert(chat.clone(), "Alice".into());
-        app.open_chat = Some(chat.clone());
-        let now = app.now();
-        app.selected_presence.select(Some(chat.clone()), now);
-        app.selected_presence.update(&chat, true, 0, now);
-
-        let mut terminal = Terminal::new(TestBackend::new(40, 8)).unwrap();
-        terminal
-            .draw(|frame| crate::ui::render_chats(frame, &mut app, Rect::new(0, 0, 40, 8)))
-            .unwrap();
-        let row = terminal
-            .backend()
-            .buffer()
-            .content()
-            .chunks(40)
-            .next()
-            .unwrap();
-        let title = row.iter().map(|cell| cell.symbol()).collect::<String>();
-        assert!(title.contains("● Alice"), "rendered title: {title:?}");
-    }
-
-    #[test]
-    fn status_broadcast_chat_never_appears_in_the_chat_list() {
-        let mut app = TestApp::new();
-        let chat = wr::JID::from("chat@example.test".to_owned());
-        let alice = wr::JID::from("alice@s.whatsapp.net".to_owned());
-        app.add_message(message(&chat, "c1", 100));
-        let mut status = message(&alice, "s1", 200);
-        status.info.chat = wr::JID::from(STATUS_BROADCAST_CHAT.to_owned());
-        app.add_message(status);
-        app.sort_chats();
-
-        assert!(app.sorted_chats.contains(&chat));
-        assert!(
-            !app.sorted_chats
-                .iter()
-                .any(|jid| jid.0.as_ref() == STATUS_BROADCAST_CHAT)
-        );
-    }
-}
